refactor: address review comments
