+++ data_types/Cargo.toml
@@ -8,7 +8,6 @@ readme = "README.md"
 
 [dependencies] # In alphabetical order
 chrono = { version = "0.4", features = ["serde"] }
-datafusion = { path = "../datafusion" }
 influxdb_line_protocol = { path = "../influxdb_line_protocol" }
 percent-encoding = "2.1.0"
 regex = "1.4"
